reorganize the plots now
