+++ python/dqmtools/dqmplots/__init__.py
@@ -1,0 +1,2 @@
+from .wibeth_plots import *
+from .daphne_plots import *
